+++ samples/Demo/Beef.Demo.Database/Migrations/20201229-114523-create-Legacy-Contact.sql
@@ -10,7 +10,6 @@ CREATE TABLE [Legacy].[Contact] (
   [Active]       BIT            NULL,
   [DontCallList] BIT            NULL,
   [AddressId]    INT            NULL,
-  [IsDeleted]    BIT            NULL, 
   CONSTRAINT [PK_Contact] PRIMARY KEY CLUSTERED ([ContactId] ASC),
   CONSTRAINT [FK_Contact_Address] FOREIGN KEY ([AddressId]) REFERENCES [Legacy].[Address] ([Id])
 );